Enhance unit test and remove workspace.xml
Improved structure of unit tests for baseEncoder. This should server
as a better example of howto.

Fixed bug in baseEncoded where it was not rounding the base value
to integer, This caused the unit test to faile for bases with decimal
values (31.5, 64.5).

Removed workspace.xml from git. This should not have been part
 of the repository. There may be more changes like this as I learn
 better what should/shouldnt go into git.